backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
tortoise-orm>=0.20.0
aerich>=0.7.2
asyncpg>=0.29.0
aiosqlite>=0.19.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
email-validator>=2.1.0
httpx>=0.27.0
apscheduler>=3.10.0
